[FIX] Pass pointer_width to code gen functions
Pass `pointer_width` to Python/C++ code generation functions.

Prior to this commit, `pointer_width` was being used to compute
structure sizes correctly, *but* generated code did *not* use the
correct pointer width.  Instead, the default pointer width of 16 was
(erroneously) used in generated code.  As of this commit, the generated
code uses the correct `pointer_width` value.

The effect of the original bug was that the data pointer for any array
argument for an RPC method would point to an undefined location in
memory.  At best, this would result in unexpected data being read.  At
worst, a segmentation fault would cause a fatal exception.