feat(i2c-bootloader): retry failed page writes 10x
Prior to this commit, if a page write failed while writing firmware to
flash memory, an exception was raised immeidately.  This approach is
problematic, as it leaves the flash memory in a non-deterministic state
which may prevent, for example, returning control to the bootloader.

This commit modifies the `TwiBootloader.write_firmware` method to retry
failed page writes up to 10 times, increasing the delay between operations
exponentially from one attempt to the next.