fix(connect): explicit port w/multiple devices
Prior to this commit, if multiple devices were found with the same device
name `SerialProxyMixin._connect` would raise `MultipleDevicesFound`, even
if a single port was explictly specified.

As of this commit, if a single port was explictly specified the connection
will be attempted.  This is useful, for example, for calling code to
explicitly select a port when encountering a `MultipleDevicesFound`
exception.