fix(async/available_devices): catch serial errors
Prior to this commit, if a serial port did not respond to an `ID_REQUEST`
request in the `available_devices()` function, the following message was
logged at the `logging.ERROR` level:

```
[ERROR:trollius]: Future/Task exception was never retrieved

WindowsError: [Error 6] The handle is invalid
```

However, an unresponsive serial port is **not _necessarily_ an error**,
since a) a device may not be connected to the port; or b) the device may
not implement the `base-node-rpc` `ID_REQUEST` protocol.

As of this commit, catch such exceptions to suppress the error messages.

Fixes #18.