fix(device-id): discard non-ID_RESPONSE packets
Prior to this commit, `_read_device_id()` assumes the next incoming packet
after writing a `ID_REQUEST` packet is a `ID_RESPONSE` packet.  However,
this may not be true if the device is writing, e.g., streaming packets
(i.e., packets with `STREAM` type).

In this commit, discard any incoming packets in `_read_device_id()` that
are not of type `ID_RESPONSE`.

Fixes #20.