fix(proxy): use 25 ms as settling_time_s default
Prior to this commit, the timeout argument of the
`base_node_rpc.async.read_device_id()` function was passed to the
top-level `asyncio.wait(..., timeout=...)` call directly (which wraps
the calls `_read_device_id()` coroutine). The default timeout of 0
caused the wrapped `_read_device_id()` coroutine call to exit before
having a chance to read a response.

In this commit, set the default `settling_time_s` timeout value to 25 ms
in `SerialProxyMixin.___init__()`. This should provide enough of a
"boot-up" delay to ensure backwards compatibility with
`base-node-rpc<0.51.1` and _at least_ Teensy devices.

Fixes #21.